requirements.txt
flask>=3.0.0
pypdf>=4.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
transformers>=4.40.0
accelerate>=0.29.0
bitsandbytes>=0.43.0
torch>=2.2.0
huggingface_hub>=0.22.0
numpy>=1.24.0
gunicorn>=21.2.0
